+++ backend/backend.go
@@ -192,6 +192,7 @@ func Initialize(config *Config) (*Backend, error) {
 		EventStore:      eventStoreProxy,
 		Bus:             bus,
 		LivenessFactory: liveness.EtcdFactory(b.ctx, b.Client),
+		Client:          b.Client,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error initializing %s: %s", event.Name(), err)
@@ -206,6 +207,7 @@ func Initialize(config *Config) (*Backend, error) {
 		Bus:         bus,
 		QueueGetter: queueGetter,
 		RingPool:    ringPool,
+		Client:      b.Client,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error initializing %s: %s", scheduler.Name(), err)