+++ backend/eventd/silenced.go
@@ -4,8 +4,9 @@ import (
 	"context"
 	"fmt"
 
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/types"
+	"github.com/sensu/sensu-go/backend/store/cache"
 	stringsutil "github.com/sensu/sensu-go/util/strings"
 )
 
@@ -21,39 +22,27 @@ func addToSilencedBy(id string, ids []string) []string {
 // getSilenced retrieves all silenced entries for a given event, using the
 // entity subscription, the check subscription and the check name while
 // supporting wildcard silenced entries (e.g. subscription:*)
-func getSilenced(ctx context.Context, event *types.Event, s store.Store) error {
+func getSilenced(ctx context.Context, event *corev2.Event, cache *cache.Resource) {
 	if !event.HasCheck() {
-		return nil
+		return
 	}
 
-	// Retrieve silenced entries using the entity subscription
-	entitySubscription := types.GetEntitySubscription(event.Entity.Name)
-	subscriptions := append(event.Check.Subscriptions, entitySubscription)
-
-	entries, err := s.GetSilencedEntriesBySubscription(ctx, subscriptions...)
-	if err != nil {
-		return fmt.Errorf("error setting silenced entries: %s", err)
+	resources := cache.Get(event.Check.Namespace)
+	entries := make([]*corev2.Silenced, len(resources))
+	for i, resource := range resources {
+		entries[i] = resource.Resource.(*corev2.Silenced)
 	}
-
-	// Retrieve silenced entries using the check name
-	results, err := s.GetSilencedEntriesByCheckName(ctx, event.Check.Name)
-	if err != nil {
-		return err
-	}
-	entries = append(entries, results...)
 
 	// Determine which entries silence this event
 	silencedIDs := silencedBy(event, entries)
 
 	// Add to the event all silenced entries ID that actually silence it
 	event.Check.Silenced = silencedIDs
-
-	return nil
 }
 
 // silencedBy determines which of the given silenced entries silenced a given
 // event and return a list of silenced entry IDs
-func silencedBy(event *types.Event, silencedEntries []*types.Silenced) []string {
+func silencedBy(event *corev2.Event, silencedEntries []*corev2.Silenced) []string {
 	silencedBy := event.SilencedBy(silencedEntries)
 	names := make([]string, 0, len(silencedBy))
 	for _, entry := range silencedBy {
@@ -62,7 +51,7 @@ func silencedBy(event *types.Event, silencedEntries []*types.Silenced) []string 
 	return names
 }
 
-func handleExpireOnResolveEntries(ctx context.Context, event *types.Event, store store.Store) error {
+func handleExpireOnResolveEntries(ctx context.Context, event *corev2.Event, store store.Store) error {
 	// Make sure we have a check and that the event is a resolution
 	if !event.HasCheck() || !event.IsResolution() {
 		return nil