+++ backend/schedulerd/check_watcher_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/backend/store/cache"
 	"github.com/sensu/sensu-go/testing/mockstore"
 	"github.com/sensu/sensu-go/types"
 	"github.com/stretchr/testify/mock"
@@ -36,7 +37,7 @@ func TestCheckWatcherSmoke(t *testing.T) {
 	watcherChan := make(chan store.WatchEventCheckConfig)
 	st.On("GetCheckConfigWatcher", mock.Anything).Return((<-chan store.WatchEventCheckConfig)(watcherChan), nil)
 
-	watcher := NewCheckWatcher(ctx, bus, st, nil, &EntityCache{})
+	watcher := NewCheckWatcher(ctx, bus, st, nil, &cache.Resource{})
 	require.NoError(t, watcher.Start())
 
 	checkAA := types.FixtureCheckConfig("a")