+++ backend/store/cache/cache_integration_test.go
@@ -1,17 +1,27 @@
 // +build integration
 
-package schedulerd
+package cache
 
 import (
 	"context"
 	"fmt"
 	"testing"
 
+	"github.com/echlebek/crock"
+	time "github.com/echlebek/timeproxy"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/etcd"
 	store "github.com/sensu/sensu-go/backend/store/etcd"
 	"github.com/sensu/sensu-go/types"
 )
+
+var mockTime = crock.NewTime(time.Unix(0, 0))
+
+func init() {
+	mockTime.Resolution = time.Millisecond
+	mockTime.Multiplier = 100
+	time.TimeProxy = mockTime
+}
 
 func TestEntityCacheIntegration(t *testing.T) {
 	mockTime.Start()
@@ -30,20 +40,20 @@ func TestEntityCacheIntegration(t *testing.T) {
 		t.Fatal(err)
 	}
 	ctx := context.WithValue(context.Background(), corev2.NamespaceKey, "default")
-	fixtures := []EntityCacheValue{}
+	fixtures := []Value{}
 
 	// Populate store with some initial entities
 	for i := 0; i < 9; i++ {
 		fixture := corev2.FixtureEntity(fmt.Sprintf("%d", i))
 		fixture.Name = fmt.Sprintf("%d", i)
 		fixture.EntityClass = corev2.EntityProxyClass
-		fixtures = append(fixtures, getEntityCacheValue(fixture))
+		fixtures = append(fixtures, getCacheValue(fixture, true))
 		if err := store.UpdateEntity(ctx, fixture); err != nil {
 			t.Fatal(err)
 		}
 	}
 
-	otherFixtures := []EntityCacheValue{}
+	otherFixtures := []Value{}
 
 	// Include some entities from a non-default namespace
 	if err := store.CreateNamespace(context.Background(), &corev2.Namespace{Name: "other"}); err != nil {
@@ -54,7 +64,7 @@ func TestEntityCacheIntegration(t *testing.T) {
 		fixture.Name = fmt.Sprintf("%d", i)
 		fixture.Namespace = "other"
 		fixture.EntityClass = corev2.EntityProxyClass
-		otherFixtures = append(otherFixtures, getEntityCacheValue(fixture))
+		otherFixtures = append(otherFixtures, getCacheValue(fixture, true))
 		if err := store.UpdateEntity(ctx, fixture); err != nil {
 			t.Fatal(err)
 		}
@@ -63,22 +73,22 @@ func TestEntityCacheIntegration(t *testing.T) {
 	cacheCtx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	cache, err := NewEntityCache(cacheCtx, store)
+	cache, err := New(cacheCtx, client, &corev2.Entity{}, true)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	watcher := cache.Watch(cacheCtx)
 
-	if got, want := cache.GetEntities("default"), fixtures; !checkEntities(got, want) {
+	if got, want := cache.Get("default"), fixtures; !checkEntities(got, want) {
 		t.Fatalf("bad entities")
 	}
 
-	if got, want := cache.GetEntities("notdefault"), []EntityCacheValue{}; !checkEntities(got, want) {
+	if got, want := cache.Get("notdefault"), []Value{}; !checkEntities(got, want) {
 		t.Fatal("bad entities")
 	}
 
-	if got, want := cache.GetEntities("other"), otherFixtures; !checkEntities(got, want) {
+	if got, want := cache.Get("other"), otherFixtures; !checkEntities(got, want) {
 		t.Fatal("bad entities")
 	}
 
@@ -90,10 +100,10 @@ func TestEntityCacheIntegration(t *testing.T) {
 
 	<-watcher
 
-	got := cache.GetEntities("default")
+	got := cache.Get("default")
 
-	if got, want := got[len(got)-1], getEntityCacheValue(newEntity); got.Entity.Name != want.Entity.Name {
-		t.Errorf("bad entity: got %s, want %s", got.Entity.Name, want.Entity.Name)
+	if got, want := got[len(got)-1], getCacheValue(newEntity, true); got.Resource.GetObjectMeta().Name != want.Resource.GetObjectMeta().Name {
+		t.Errorf("bad entity: got %s, want %s", got.Resource.GetObjectMeta().Name, want.Resource.GetObjectMeta().Name)
 	}
 
 	if err := store.DeleteEntity(ctx, newEntity); err != nil {
@@ -102,21 +112,21 @@ func TestEntityCacheIntegration(t *testing.T) {
 
 	<-watcher
 
-	if got, want := cache.GetEntities("default"), fixtures; !checkEntities(got, want) {
+	if got, want := cache.Get("default"), fixtures; !checkEntities(got, want) {
 		t.Errorf("bad entities")
 	}
 
 }
 
-func checkEntities(got, want []EntityCacheValue) bool {
+func checkEntities(got, want []Value) bool {
 	if len(got) != len(want) {
 		return false
 	}
 	for i := range got {
-		if got[i].Entity.Namespace != want[i].Entity.Namespace {
+		if got[i].Resource.GetObjectMeta().Namespace != want[i].Resource.GetObjectMeta().Namespace {
 			return false
 		}
-		if got[i].Entity.Name != want[i].Entity.Name {
+		if got[i].Resource.GetObjectMeta().Name != want[i].Resource.GetObjectMeta().Name {
 			return false
 		}
 	}