+++ backend/store/cache/cache_test.go
@@ -1,0 +1,135 @@
+package cache
+
+import (
+	"reflect"
+	"testing"
+
+	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/types/dynamic"
+
+	"github.com/sensu/sensu-go/testing/fixture"
+	"github.com/stretchr/testify/assert"
+
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
+)
+
+func fixtureEntity(namespace, name string) *corev2.Entity {
+	entity := corev2.FixtureEntity(name)
+	entity.Namespace = namespace
+	return entity
+}
+
+func TestCacheGet(t *testing.T) {
+	cache := Resource{
+		cache: buildCache([]corev2.Resource{
+			fixtureEntity("a", "1"),
+			fixtureEntity("a", "2"),
+			fixtureEntity("a", "3"),
+			fixtureEntity("a", "4"),
+			fixtureEntity("a", "5"),
+			fixtureEntity("a", "6"),
+			fixtureEntity("b", "1"),
+			fixtureEntity("b", "2"),
+			fixtureEntity("b", "3"),
+			fixtureEntity("b", "4"),
+			fixtureEntity("b", "5"),
+			fixtureEntity("b", "6"),
+			fixtureEntity("c", "1"),
+			fixtureEntity("c", "2"),
+			fixtureEntity("c", "3"),
+			fixtureEntity("c", "4"),
+			fixtureEntity("c", "5"),
+			fixtureEntity("c", "6"),
+		},
+			true,
+		),
+	}
+	want := []Value{
+		{Resource: fixtureEntity("b", "1"), Synth: dynamic.Synthesize(fixtureEntity("b", "1"))},
+		{Resource: fixtureEntity("b", "2"), Synth: dynamic.Synthesize(fixtureEntity("b", "2"))},
+		{Resource: fixtureEntity("b", "3"), Synth: dynamic.Synthesize(fixtureEntity("b", "3"))},
+		{Resource: fixtureEntity("b", "4"), Synth: dynamic.Synthesize(fixtureEntity("b", "4"))},
+		{Resource: fixtureEntity("b", "5"), Synth: dynamic.Synthesize(fixtureEntity("b", "5"))},
+		{Resource: fixtureEntity("b", "6"), Synth: dynamic.Synthesize(fixtureEntity("b", "6"))},
+	}
+	got := cache.Get("b")
+	if !reflect.DeepEqual(got, want) {
+		t.Fatalf("bad resources: got %v, want %v", got, want)
+	}
+}
+
+func TestBuildCache(t *testing.T) {
+	resource1 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource1", Namespace: "default"}}
+	resource2 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource2", Namespace: "default"}}
+	resource3 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource3", Namespace: "acme"}}
+
+	cache := buildCache([]corev2.Resource{resource1, resource2, resource3}, false)
+
+	assert.Len(t, cache["acme"], 1)
+	assert.Len(t, cache["default"], 2)
+	assert.Len(t, cache, 2)
+}
+
+func TestResourceUpdateCache(t *testing.T) {
+	cacher := Resource{cache: make(map[string][]Value)}
+	resource0 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource0", Namespace: "default"}, Foo: "bar"}
+	resource1 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource1", Namespace: "default"}, Foo: "bar"}
+	resource2 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource0", Namespace: "default"}, Foo: "baz"}
+	resource3 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource1", Namespace: "default"}, Foo: "qux"}
+
+	// Add a resource
+	cacher.updates = append(cacher.updates, store.WatchEventResource{
+		Resource: resource1,
+		Action:   store.WatchCreate,
+	})
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 1)
+
+	// Add a second resource. It should be alphabetically sorted and therefore at
+	// the beginning of the namespace cache values even if it was appended at the
+	// end
+	cacher.updates = append(cacher.updates, store.WatchEventResource{
+		Resource: resource0, Action: store.WatchCreate,
+	})
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 2)
+	assert.Equal(t, resource0, cacher.cache["default"][0].Resource)
+	assert.Equal(t, resource1, cacher.cache["default"][1].Resource)
+
+	// Update the resources
+	updates := []store.WatchEventResource{
+		store.WatchEventResource{Resource: resource2, Action: store.WatchUpdate},
+		store.WatchEventResource{Resource: resource3, Action: store.WatchUpdate},
+	}
+	cacher.updates = append(cacher.updates, updates...)
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 2)
+	assert.Equal(t, resource2, cacher.cache["default"][0].Resource.(*fixture.Resource))
+	assert.Equal(t, resource3, cacher.cache["default"][1].Resource.(*fixture.Resource))
+
+	// Delete the resources
+	deletes := []store.WatchEventResource{
+		store.WatchEventResource{Resource: resource3, Action: store.WatchDelete},
+		store.WatchEventResource{Resource: resource2, Action: store.WatchDelete},
+	}
+	cacher.updates = append(cacher.updates, deletes...)
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 0)
+
+	// Invalid watch event
+	var nilResource *fixture.Resource
+	cacher.updates = append(cacher.updates, store.WatchEventResource{
+		Resource: nilResource,
+		Action:   store.WatchCreate,
+	})
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 0)
+
+	// Unknown watch event type
+	cacher.updates = append(cacher.updates, store.WatchEventResource{
+		Resource: resource0,
+		Action:   store.WatchUnknown,
+	})
+	cacher.updateCache()
+	assert.Len(t, cacher.cache["default"], 0)
+}