+++ backend/store/cache/logger.go
@@ -1,0 +1,7 @@
+package cache
+
+import "github.com/sirupsen/logrus"
+
+var logger = logrus.WithFields(logrus.Fields{
+	"component": "cache",
+})