+++ testing/mockstore/entities.go
@@ -36,9 +36,3 @@ func (s *MockStore) UpdateEntity(ctx context.Context, e *types.Entity) error {
 	args := s.Called(ctx, e)
 	return args.Error(0)
 }
-
-// GetEntityWatcher
-func (s *MockStore) GetEntityWatcher(ctx context.Context) <-chan store.WatchEventEntity {
-	args := s.Called(ctx)
-	return args.Get(0).(<-chan store.WatchEventEntity)
-}